+++ test/schema.test.ts
@@ -1,8 +1,17 @@
-import { User } from './models/user';
+import { Friend, User } from './models/user';
 import { a, getSchemaByType, validate } from '../src/index';
 
 describe('EnYupSchema method ', () => {
-  describe('preserve class instances', function() {
+  describe('preserve class instances', function () {
+    it('should convert valid objects in instances', async () => {
+      const actual = await validate({
+        object: getValidObject(),
+        schemaName: 'user',
+      });
+      expect(actual).toEqual(getValidUser());
+      expect(actual instanceof User).toBe(true);
+    });
+
     it('should convert valid objects in instances', async () => {
       const actual = await validate({
         object: getValidObject(),
@@ -55,7 +64,7 @@ describe('EnYupSchema method ', () => {
   it('should omit', async () => {
     expect(
       getSchemaByType(User)
-        .omit(['name'])
+        .omit(['name', 'friends'])
         .describe()
     ).toEqual(a.object({ birthday: a.date().required() }).describe());
   });
@@ -64,9 +73,9 @@ describe('EnYupSchema method ', () => {
 const birthday = new Date(1997, 11, 12);
 
 function getValidUser() {
-  return new User({ name: 'Mattia', birthday });
+  return new User({ name: 'Mattia', birthday, friends: { '1': new Friend({ name: "Vincenzo", }) } });
 }
 
 function getValidObject() {
-  return { name: 'Mattia', birthday: birthday.toString() };
+  return { name: 'Mattia', birthday: birthday.toString(), friends: { '1': { name: "Vincenzo" } } };
 }