+++ CHANGELOG.md
@@ -1,3 +1,13 @@
+# UNRELEASED
+### Added
+
+- `nestedObject` to register an object where each value is of the given property
+- `@is` accept yup.lazy
+
+### Changed
+
+- more documentation
+
 # 1.1.0 (2025-01-04)
 
 ### Added